Change graph indices to point at start of mer
This change essentially undoes the logic of the kmer-indexed-at-end
approach. We now point at the the front of the kmer, and during
alignment our only guarantee is that there is one allele that will have
perfect alignment. I’m making this choice because at the moment, I
don’t have the right primitives in alignment to accurately propagate
the mismatch calculation to the end of the kmer.