Use previous positions adjacency calc as seed.
Seems like a no-brainer given the current compute all the positions
approach. It also seems to make some positions in DRB1 actually
"tractable" as some positions result in calculations on the order
of minutes (Not 100% sure that another bug isn't lurking), while
this new sequential method takes a little bit less than 4 minutes
for all the positions in DRB1.