Correct the FoldAtSamePosition to account for S/E nodes.
Start and End nodes are at the same positions as their successors
and predecessors, this make it difficult to fold over all of the nodes
at given position. But since we initialize the fold with all of the
start nodes and end nodes have no successor we can exclude them from
the add_successor check.