Use custom fixed width bitvectors for allele sets
This PR addresses the current major hotspot in the ParPHMM computation:
allele union and difference passes. We use a custom bitvector
implementation, in Fixed_width, that is as simple as possible but not
properly abstracted. The user MUST specify the length before using the
bitvectors, or out of bounds errors will surface. The resulting code
reduces the per-read running time by about 40%!