Unify the insert, transition reference size logic.
Insert probabilities are now configurable and use a default value that
is interpreted the same way in the single and multiple allele mode.

Forward pass dynamic array size is now determined, unless specified, by
the actual data structure. The use of the reference size is now also
something that may be overwritten, as it is for the single allele case,
when determining the transition probabilities.