Parameterize the CAM merging operations on cell equality.
The forward pass of the PHMM bifurcates as the forward recurrences
compute different values. Sufficiently often, these values are only
different by some small values (ie < 1e-8). Unfortunately, each
bifurcation causes another element to be storred in the CAM and
consequently raises the running time of a merge. To eliminate these
extra values's we'll use a lower resolution float accuracy to determine
if two cells are equal. Note that this requires more float comparisons,
so the reduction in accuracy must yield sufficiently fewer cells inorder
for this optimization to be worthwile.