Investigate and implement CAM paring lookups.
While we perform various maps and concat_maps in CAM we can trim the
intersections that we've correctly aligned. This requires recreating the
CAM as we perform the lookup and therein lies the rub. That can be more
expensive. This commit leaves commented out original and possible future
logic, but is at the moment optimized for our use cases. The code is
left as is to allow a reboot of the context of this work as I can't
really see any other way forward along this path.