Add simple, stupid, merging logic to merge4.
Everytime that we have a new value to store in merge4 scan the
accumulator and merge the new intersect with any previous sets if their
values are close-enough numerically. Surprisingly, this simple, stupid,
solution is much faster than other solutions with pre-allocated storage
arrays; these are hard to get right for different reads.