Encode the pair as an int and simplify interdiff.
We assume a 64bit system and encode the start and end of the intervals
of a partition in the top 31 and bottom 32bits of an int. This leads to
several improvements in how we check for intersections and lets us get
rid of the messy intersection-difference function and the allocations
that it required. The resulting code is now 30% faster, because we've
reduced the allocation burden.