Embed the position of mostly likely emission in Pass output.
We need this position to form a coverage model, this makes it a pretty
fundamental output of a forward pass. Therefore I'm embedding it at
almost the bottom of the logic. This PR also simplifies the position and
likelihood tracking that results from having these two data points
returned jointly.